+++ pytsite/auth_http_api/__init__.py
@@ -25,10 +25,10 @@ def _init():
     http_api.handle('PATCH', 'auth/user/<uid>', _controllers.PatchUser(), 'pytsite.auth@patch_user')
 
     # Following HTTP API
-    http_api.handle('POST', 'auth/follow/<uid>', _controllers.PostFollow(), 'pytsite.auth@post_follow')
-    http_api.handle('DELETE', 'auth/follow/<uid>', _controllers.DeleteFollow(), 'pytsite.auth@delete_follow')
     http_api.handle('GET', 'auth/follows/<uid>', _controllers.GetFollowsOrFollowers(), 'pytsite.auth@get_follows')
     http_api.handle('GET', 'auth/followers/<uid>', _controllers.GetFollowsOrFollowers(), 'pytsite.auth@get_followers')
+    http_api.handle('POST', 'auth/follow/<uid>', _controllers.PostFollow(), 'pytsite.auth@post_follow')
+    http_api.handle('DELETE', 'auth/follow/<uid>', _controllers.DeleteFollow(), 'pytsite.auth@delete_follow')
 
     # Block users HTTP API
     http_api.handle('POST', 'auth/block_user/<uid>', _controllers.PostBlockUser(), 'pytsite.auth@post_block_user')